Refactor: Lotto클래스로 로또 번호 중복 로직 역할 이동, 랜덤 로직 리팩토링, 로또 번호 반환 자료구조 변경
