+++ src/main/kotlin/lotto/data/LottoNumber.kt
@@ -15,25 +15,16 @@ class LottoNumber private constructor(private val number: Int) : Comparable<Lott
     companion object {
         private const val MIN_NUMBER = 1
         private const val MAX_NUMBER = 45
-        private const val LOTTO_NUMBER_LENGTH = 6
         private const val ERR_MSG_OUT_OF_LOTTO_RANGE = "로또의 범위를 넘어서는 번호입니다."
-        private const val ERR_MSG_OUT_OF_LOTTO_LENGTH = "번호는 6개로 구성되어야 합니다."
 
         private val NUMBERS: Map<Int, LottoNumber> = (MIN_NUMBER..MAX_NUMBER).associateWith(::LottoNumber)
 
-        fun createLottoNumbers(numbers: List<Int>): LinkedHashSet<LottoNumber> {
-            val lottoNumbers = LinkedHashSet(numbers.map(::from))
-            validateDuplication(lottoNumbers)
-
-            return lottoNumbers
+        fun createLottoNumbers(numbers: List<Int>): Set<LottoNumber> {
+            return LinkedHashSet(numbers.map(::from))
         }
 
-        fun createRandomLotto(randomLogic: RandomLogicInterface): LinkedHashSet<LottoNumber> {
-            return randomLogic.createRandomLotto(NUMBERS)
-        }
-
-        private fun validateDuplication(lottoNumbers: LinkedHashSet<LottoNumber>) {
-            require(lottoNumbers.size == LOTTO_NUMBER_LENGTH) { ERR_MSG_OUT_OF_LOTTO_LENGTH }
+        fun createRandomLottoNumber(lottoCreation: RandomLogicInterface): Set<LottoNumber> {
+            return lottoCreation.createAutoLotto(NUMBERS)
         }
 
         private fun from(value: Int): LottoNumber {
