+++ src/main/kotlin/lotto/domain/RandomLogicInterface.kt
@@ -3,5 +3,5 @@ package lotto.domain
 import lotto.data.LottoNumber
 
 interface RandomLogicInterface {
-    fun createRandomLotto(numbers: Map<Int, LottoNumber>): LinkedHashSet<LottoNumber>
+    fun createAutoLotto(numbers: Map<Int, LottoNumber>): Set<LottoNumber>
 }
